[FLINK-35344][cdc-base] Move same code from multiple subclasses to JdbcSourceChunkSplitter
